requirements.txt
click==8.0.4
confuse==2.0.0
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paho-mqtt==1.6.1
PyYAML==6.0
six==1.16.0
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
PyJWT==2.4.0
